client: Accept raw string when creating admin portal
This is a breaking change that simplifies instantiating an "AdminPortal".
Now accepts a raw string which is parsed to a URL as validation.

This reduces some complexity in the creating of the HTTP requests
internally and also provides a Fixes #31 ensuring we can support
non-standard installs that are installed under a specific path.